Add the `--statistics` command line option.
Gather some simple statistics about the contacts and print them out when
the `--statistics` option is specified.

Add the `contacts` struct to hold both the array of individual `contact`
structs and the statistics: count, max emails count and max phones
count.  Add the `alloc_contacts()` function and change `free_contacts()`
to work on a `contacts` struct.  Change functions that previously took a
pointer to an array of `contact` structs plus a count to use a
`contacts` struct instead.  Change the `enumerate_contacts()` function
so that it creates the `contacts` struct instead of the bare array of
`contact` structs.